Deleting the category from your page now just removes you instead of the entire category

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
 
@@ -104,7 +105,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
+        $category->users()->detach(Auth::user());
+        if($category->users->isEmpty()) {
+            $category->delete();
+        }
         return redirect(route('user.show'))->with('message', 'Category deleted successfully');
     }
 }
